create display generation results command
